\name{curvHDR-internal}
\title{Internal curvHDR Objects}
\alias{addIntervs}
\alias{areaPolygon}
\alias{contour1D}
\alias{intervalGrow}
\alias{plot3D}
\alias{polygonGrow}
\alias{polyhedronGrow}
\alias{ptsInTriPolyh}
\alias{rgl.lines3d}
\alias{separateTriPolyh}
\alias{tranUnitInt}
\alias{triPolyhCentroid}
\description{Internal curvHDR functions.}
\details{These are not to be called by the user.}
\keyword{internal}
