\name{curvHDRvignette}
\alias{curvHDRvignette}
\title{Display the package's vignette.}
\description{The vignette of the curvHDR package is displayed using the default PDF file browser. It provides a detailed detailed description of use of the package for gating flow cytometry data using the curvHDR method.}
\usage{
curvHDRvignette()
}
\author{Matt Wand\email{matt.wand@uts.edu.au}, G. Luta\email{gl77@georgetown.edu} and U. Naumann\email{ulrike.naumann1@gmail.com}}
\examples{
curvHDRvignette()
}

